tests for findUncertainPair

+++ dedupe/api.py
@@ -735,7 +735,7 @@ class ActiveMatching(Matching):
 
     def _trainClassifier(self):  # pragma : no cover
         labels = numpy.array(self.training_data['label'] == b'match',
-                             dtype='int32')
+                             dtype='int8')
         examples = self.training_data['distances']
 
         self.classifier.fit(examples, labels)
